requirements.txt
streamlit
pandas
numpy
openai
watchdog
python-dotenv
folium
streamlit-folium
